Refactoring
Moves and renames 'StoryStruct' to the more appropriate 'CustomStaticCache' and related refactoring
